# reduce memory footprint of json.Marshal for large values