@pelletier 2019-12-04
For those who don't have Slack access: this change makes `Marshal` use a buffer pool instead of always allocating a new 1KiB array each time.